requirements.txt
# pvc general imports
pyopengl
numpy

# pvc for windows
SpoutGL>=0.0.4; platform_system == "Windows"

# pvc for macos
syphonpy; platform_system == "Darwin"
glfw; platform_system == "Darwin"